Fix Errors
I hardcoded the Schwarzschild radius and did not allow it to be changed in a function call, as well as an issue in my energy equation where the square root of the equation was not including the
Schwarzschild metric. Which made the affect of light bending more drastic as you approached the center of the black hole.

+++ Main.cpp
@@ -19,7 +19,7 @@ void onResize(GLFWwindow*, int w, int h)
 }
 
 // Pixel Buffer: setting each pixel:
-void setPixel(float x, float y, float red, float green, float blue)
+void setPixel(int x, int y, float red, float green, float blue)
 {
     int index = (y * WIDTH + x) * 3;
     framebuffer[index + 0] = red;
